Store items as `String` in cw-core.
This makes items a little more flexible and performant. Before we
limited items to being `Addr` types, but this meant that if you wanted
to do something as simple as customize the color of your DAO page you
would need to create a new contract for that and the frontend would
need to make two queries.

This also makes the `list_items` query return key-value pairs so that
the frontend doesn't need to make a second query to get the value of
an item it has determined exists.